Yönlendirmedeki hatalar giderildi

Related work items: #683

+++ src/pages/surveys/TumAnketler.jsx
@@ -3,34 +3,43 @@ import Table from "../../components/Table/Table";
 import Layout from "../../components/Layout";
 import SurveyService from "../../services/SurveyService";
 import { Link, useNavigate } from "react-router-dom";
-import Alert from "../../components/Alert";
 import BreadCrumbs from "../../components/BreadCrumbs";
 
 const TumAnketler = () => {
   const [surveys, setSurveys] = useState([]);
 
   const header = ["No", "Anket Adı", "Konu Başlığı"];
-
+  const [deleteSurvey, setDeleteSurvey] = useState(false);
   const deleteTableRows = async (index, rowData) => {
-    const response = await SurveyService.delete(rowData.surveyOid);
-    if (response.status === 200) {
-      alert(rowData.surveyOid + ". id'ye sahip anket başarıyla silindi");
-    } else {
-      alert("bir hata meydana geldi");
+    const shouldDelete = window.confirm("Bu anketi silmek istediğinize emin misiniz?");
+    if (shouldDelete) {
+      try {
+        const response = await SurveyService.delete(rowData.surveyOid);
+        console.log(response);
+        if (response.status === 200) {
+          alert(rowData.surveyOid + ". id'ye sahip anket başarıyla silindi");
+          const rows = [...surveys];
+          rows.splice(index, 1);
+          setSurveys(rows);
+          setDeleteSurvey(true);
+        } else {
+          alert("Bir hata meydana geldi");
+        }
+      } catch (error) {
+        console.log(error);
+        alert("Bir hata meydana geldi");
+      }
     }
-    console.log(response);
-    const rows = [...surveys];
-    rows.splice(index, 1);
-    setSurveys(rows);
   };
-
   const navigate = useNavigate();
 
   const handleEditClick = (rowData) => {
+    console.log(rowData);
     navigate(`/anketler/guncelle/${rowData.surveyOid}`, { state: rowData });
   };
 
   useEffect(() => {
+    setDeleteSurvey(false);
     const fetchSurveys = async () => {
       try {
         const response = await SurveyService.list();
@@ -55,7 +64,7 @@ const TumAnketler = () => {
     return () => {
       console.log("useEffect clean up");
     };
-  }, []);
+  }, [deleteSurvey]);
 
   const header2 = { header: "Tüm Anketler", href: "/anketler" };
